Music
As a group, we should add more songs and choose one. It may be too hard to find a banjo part, as I could not find very many modern songs that were for a four-stringed banjo or only have four notes. We also debated using a ukulele for a third instrument.
We planned to have two songs for each instrument, including acoustic and electric guitars. We'll see if this is too much to handle; the first step is to choose a song and find recordings of each of the chords for it.