Keep Processing_Hero_0 Cleaner
Please next time only put the functions of your class in there. Anything
like the fill or strokeWeight should be in the class's tab.
This would make it so much easier to add the if boolean and if
mousepressed controls rather than having to sort through what should go
where.
Also: do we need to add how we expect our game to work into the
flowchart rather than all that is literally in the master branch
currently?
Also: we need to make GameOver change depending on score, and we need to
make it work.
It's also really slow.