It's Alive
It could merge after I told it a joke.
It. Merged.
I'm gonna cry. I also did a bit of editing to clean it up. I'll delete
some files next time.
This is has been a roller coaster ride of a day. And I hate roller
coasters.
But to appease the app which apparently appreciates hilarity, here's
another gem:
We're gits of the round table! | We eat ham and jam and commitalot! | On
second thought, let's not go to #gitalot. It is a silly place.