Whoopie
Cushion. Pie. ing Cough.  I'm so happy that it's working now for some
unknown (to me) reason. However, there's still work to do. This
includes:
Minim. Someone figure this out and add the files so we can drop the
base.
Colors of the start screen. Pretty sure they were prettier before i got
my hands on em.
The GameOver Screen. I may have kinda misplaced it. I'll search for  the
code.
The Score. We need a setup. And I'll add in character, song, and guitar
choosing screens. Plus the instructions/Pause overlay during game play.
But I'm still happy with how this is working out. Guess you could say
that this really is the git that keeps on giving.