Added Files
Bit more work on the code, but still not functional. Finished character
album pictures for character selection, and added files showing my work
on all the graphics that I've done. I'll remove them later to clean up,
I just want a place to keep them for now.
And another git joke because I love the repo I'm finding them on:
May the forks be with you.