Screens
Clicking now works. Finalized guitar screen, including extra guitar.
Issues still with the music selection screen going into the game play
screen, but it's getting there. Had to readd a BUNCH of pictures because
the cropping and white background on them made it difficult to work with
and difficult to look at, respectively.