EMMA ISSUES 4 UUUU
Go back to the other one and look for all comments tagged Emma. Just try
playing this; it doesn't work. So I'll work on it. Then we can add the
song.
So now we just need to choose the second acoustic song (something you
guys would like that wouldn't be too simple). See facebook.