I forgot how much I messed this up
If someone could help out:
Emma, go through "Screens" and "EMMA ISSUES 4 UUUU" and mark down all
the times I've typed up your name. I'll add the main issues from this
commit and the others if you don't get a chance to tonight.
Caitie, I'll try to fix the start screen. If I don't tonight, I'm
leaving it up to you.
Sam, if you could find the final song online and add it in, that'd also
be appreciated.
I know you guys might not see this with midterms and all on our minds,
but there is always a chance.