Song Names
The records are now labeled. The font color is red no matter what I do,
like the start screen, so that could be a hint to what's really going on
if I wanted to think about it. But I'll just focus on getting ellipses
to fall as the song plays, so if someone else could figure this out,
that'd be rad.
Thanks!