Music
I think this is the most beautiful thing I have ever created. Hit play
and really /listen/ to this mixture of sounds. By golly, it's inspiring.
I think the reason it is SO messed up is that it plays a new version of
the song each and every time the draw loop goes through it, so it
progresses like a sort of very-closely-spaced round. It is no less than
magical.
And very, very far from what I intended.