much cooler start screen
the old one is there but i turned it off because i like this one much
much more. let me know what you guys think -- i know the guitar isnt
visible at first, but i thought it was kind of cool how it slowly
appeared, like a shadow or something. if you dont like it we have a
couple options. I can do a different start screen, we can choose a
different guitar (for the "i wanna rock" button), or we can change the
color of the guitar. let me know :)