Kind of Working
The Processing_Hero_0 does not switch screens correctly. I'll try and
figure out why, but this is getting frustrating and annoying. I really
need to study for history.
The ant character is looking great, but I still need SOMEONE to make
sure the other characters and guitars can line up right. It's really
easy to do, it just takes some time. Text me if you have some time to
work on this.
Sam is getting the other acoustic song, but I need someone to make an
array list so the notes will fall. Anyone?
The Score variable and trackers still needs to be added.