Falling Notes
This was all in the game_play_test folder:
I tried making notes fall randomly, but it's not working. I need to find
away for them to be removed if the key is pressed and they are close
enough to the bottom. If anyone knows how to do this, feel free to take
a stab at it.
We need this so there is actually a game /to/ play.
I'll just work on score in the next commit, but we need to figure this
out so we'll have something to raise the score.
We need to figure out how to make the ellipses fall at different times
in their specified x value, so they fall on their line.
Please, someone take a look at the code in the Processing_Hero_0 and see
if you can get the screens sorted out.
When I added P2D to size(), the color of the lines of the start screen
changed from multi to only red, orange, and yellow colors. The numClicks
is messing it all up, but we can't just have the mousePressed and expect
it to work like that.
I really don't understand why it works and why it doesn't, so PLEASE
someone else take a look at this. I know Caitie, you have (THANK YOU),
and Sam, I'll take a look at your game_over to merge with master in a
moment, but we need a GAME.