Cleared Up
I'll merge this, but, Sam, were you planning on having notes fall?
because there was nothing in these tabs, and it's getting late.
I don't know what "Thumbs" is, but it's there.
Also, for reference, Sam is having gitHub issues and cannot unrevert her
latest commit so that we can have our second acoustic song on file.