Deleted Unnecessary
Processing_Hero_0 had nothing in it, but then I could not open the
sketch.
My computer keeps freezing. Normally, I can rely on it. But right now,
it's really slow with Processing and GitHub.
I had to rename the main folder, but it should be in order now. I also
added in all the if() statements for each boolean  and mouse location,
but it is extremely laggy or it does not work.