MousePressed
"Power co" for Caitie. By getting rid of the mousePressed functions
within the classes, the screens change in order and are displayed for
more than one frame. This is also because we added boolean oldClicks.
Thank you for your help!
Major issues still persist, though. The song selection screen's records
are not the correct sizes, and the game_play_test's minim issues cause
it not to display or play anything. This is huge. But so is our success
today.